usv_system/requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
PyYAML>=6.0
shapely>=1.8.0
cvxpy>=1.2.0
control>=0.9.0
pandas>=1.3.0
tqdm>=4.62.0
pytest>=6.2.0
pyglet==1.5.27
opencv-python==4.6.0.66
pillow==9.1.1
pyqt5==5.15.7 